requirements.txt
fastapi
uvicorn
openai
requests
transliterate
python-dotenv